chore: migrate user_matching_test into the right folder
chore: resolves https://github.com/UnlockedLabs/UnlockEdv2/pull/1165#discussion_r3405682410
